apps/worker/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
supabase==2.7.4
boto3==1.34.157
redis==5.0.8
rq==1.16.2
anthropic==0.34.2
openai==1.46.0
assemblyai==0.30.0
librosa==0.10.2
soundfile==0.12.1
ffmpeg-python==0.2.0
httpx==0.27.2
# yt-dlp installed system-wide in Docker
sentry-sdk==2.14.0
